+++ src/main/java/lunarlander/LunarLanderScore.java
@@ -10,7 +10,7 @@ public class LunarLanderScore {
         this.finalSpeed = finalSpeed;
     }
 
-    public int computeScore() {
-        return (int)((100 - finalSpeed) + 0.05*remainingFuel);
+    public double computeScore() {
+        return ((100 - finalSpeed) + 0.05*remainingFuel);
     }
 }
